FINAL VERSION 1.0

+++ database/migrations/2024_07_07_105720_create_points_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('points', function (Blueprint $table) {
             $table->id();
-            $table->integer('percentage_from_totalprice')->nullable(false);
+            $table->decimal('percentage_from_totalprice')->nullable(false);
             $table->integer('money_per_poin')->nullable(false);
             $table->timestamps();
         });
